+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Added `copyWith` member function.
+- Added `isTheSameDate` DateTime extension.
+- Added `isTheSameDate` member function.
+
 ## 1.0.0+4
 
 - Minor correction in CHANGELOG.